Combined HZC Electric vendor files
